host/requirements.txt
gTTs
pyttsx3
pygame
pyserial
